Add timestamp to label
Comparing two tests with the same label makes
it hard to understand which metrics belongs
to each run. We now add a timestamp to the
label, so when compare different runs in
WebPageTest it's easy to spot which metrics
belongs to which test.

Also avoid logging the URL for the sent
requests, the URL is wrong because the method
is async.

Bug: T113448
Change-Id: I3c204fe3e48f9db44d8bed5da3692998a119ee41