Set location by environment variable
We need to be able to test the batch scripts locally when Jenkins is
up and running (the location was hardcoded before). we now set the
location part as an environment variable.

Do it with <%WMF_WPT_LOCATION>. You can read more about it here:
https://wikitech.wikimedia.org/wiki/WebPageTest#Changing_and_testing_a_batch_file

The configuration in Jenkins has been updated so that the location is set
to east, we can safely merge this if everything looks ok.

Change-Id: Ic0ec8e617849ce07ce358527e6f1bc959225ca13