Add test runs for test2, beta and iPhone
Discussed with Ori last week what to test and
we wanted to add test for test2 and beta. However
beta is not SPDY (and https) so I haven't added
tests for logged in users on that one. And not sure
if it actually will help us as long as it isn't running
SPDY.

Merge this after https://gerrit.wikimedia.org/r/249363 so
that tests will not timeout.

Also added one run for iPhone 6. Let's keep it that way until we see
that it actually works (guess it could be quite busy). I love
having metrics for a new iOS device!

Bug: T110230
Change-Id: I1dfe4ebde653fc0082b9d88b8d42dec824b92395