Verify the base page on WPT.org
A couple of days ago we had an increase in the TTFB
on our WPT agent. To easier verify that it's true, it
good to collect metrics from another instance. So lets
verify the blank page on WPT.org.

When this reaches Jenkins, we should split the currrent
job into two, so we can run one with only the wpt.org job
(longer timeouts).

Change-Id: Ib2d44ab45848dc5e72a604e89af0d04a01620d65