Add support for reporting metrics using more than statsv
Reworked the structure so we can have different reporters
reporting metrics. We now support statsv, json, csv, and
Graphite. We can add more reporters later.

There's a new parameter --reporter that needs to be configured for
each run (fixed in the batch files) so this should work out of the
box in Jenkins. But one change is that the authenticated part of
the key has been moved to the namespace, so the graph queries need
to be changed.

Also documented the methods.

There a new bash script that should be used to verify changes in
batch files. Check out the documentation at the top of
the file: test/rulethemall.sh

Also cleaned up the default values to make it easier to find.

Bug: T114343
Change-Id: I7bd582a4a180b07b62d6c94552555957baf34ce6