Removed tcpdump and devtools
When we got this running on Jenkins the SpeedIndex and Start Render
increased for all URL:s testing using Chrome. It's a really small
chance that it is an regression, everything points to that it is
the collection of extra data that makes this happen so remove it.

Later on we could try out WPT on a larger Amazon instance.

Bug: T116981
Change-Id: Idf023783854444021c03ed1eedd18a0a64feda70