Catch when WPT misses repeat view
When WebPageTest gets some error (like error 99998) that somehow
means the test wasn't ok (even though the API returns ok) it skips
testing repeat view and we missed to catch that on our side.

Lets check that the view exists before we fetch the data and follow
up with an issue for WPT on GitHub.

Bug:T118441
Change-Id: I52bdf106d5809615cc28fa2d767bf2655aa615a6