Report errors when WebPageTest fails
Before this patch we reported OK whatever happend to the test
at WebPageTest. Now we report a fail if one of the tests in a
batch fails in WebPageTest or if WebPageTest reports a result
error code (that happen when the WPT API don't catch the error
or a request that get an 40X or 50X error).

This will make it possible to act on fails when we run the jobs
in Jenkins.

Bug: T119865
Change-Id: If4006f30fbe3087688d03bf730b1c34cbdae6814