Always use the fastest run (as median run)
When the Chrome team test a change in Chrome using WebPageTest they use
the fastest of the runs to get less noise in the numbers. We got
our PR merged in the WebPageTest API that makes that possible, so we can
also use that. Lets test it out and see if we get even better straight
lines in Grafana.

Bug: T121618
Change-Id: I33046a6961213a9c11ccbde07c05c3b1119821f5