Make it easier to see changes in videos for logged in users
The current running version logs in the user and then navigates to
the next page. That makes the videos for the last page harder to follow.
Lets remove the HTML from the login page, wait 2 seconds and the navigate
to the end point. The video will now be white in the beginning and it's
easier to see start render etc.

Waiting on our agent to updated to latest version to remove the
password bubble to make this perfect.

Bug: T125784
Change-Id: Ib2920ffbf2b815d0ca7890e079ff385f858d558c