Append the date on the label instead of prepend
It's hard sometimes to find the right run in the WebPageTest test history
since it is cut-off in the display. Today we start with the date and then
have the name of the test. Lets do it the other way around to make it
easier to find the right test.

Bug: T126959
Change-Id: Ie15ef3870eb99df54ae4a0e489f3cb2f07d4da28