Include Brasil (pt wiki) in webpagetest runs
pt is a smaller wiki. As part of the lazy loading references work
we would like to run a test on this wiki. Having this represented in
tests would thus be useful. The Brazil pt page is a good representative article
where references seems to carry a lot of weight in the page HTML.

Bug: T126802
Change-Id: I11c64b2e766680aa3c8f054d683e9cb01d6e3282