Capture Japanese wiki article in tests
This adds data for the Japanese Japan article to get a sense
of performance impact for this wiki with regards to our changes.

Page was chosen as it's HTML is large (960KB)
10% of HTML on the Japan article is references
23% of the HTML is navboxes

It also adds Barack Obama where 50% of page is references and 7%
is navboxes

We are preparing to roll out a patch to lazy load references on
Japanese wiki so this will help with that.

Bug: T126802
Change-Id: I621827173403626cae93361f4ec2313eeb945b90