Move non-working Amazon tests to wpt.org
The Brasil test runs have a strange bug: Running on our own Amazon instance
the test looks fine on wpt.wmftest.org but the JSON API somehow breaks and
return empty so we cannot send the stats to Graphite. However when we run on
webpagetest.org it works. We have filed an issue:
https://github.com/WPO-Foundation/webpagetest/issues/569

Until it is fixed, lets use webpagetest.org so we can get metrics.

Bug: T127833
Change-Id: Iec010a451f45331bb589b55e14a8e5e61db55458