Get metrics from first run only for WebPageTest.org
We have changed so we only collect metrics from the first run on our
own WebPageTest instance. Lets do the same on WebPageTest.org so we
run the tests exactly the same + have more runs to run on real mobile
phones (we have a limited amount of runs we can do per day).

Bug: T127798
Change-Id: I80a773f9c5b931ebf2ffc5ca88a888e31d8e824c