Clear the screen between runs to make SpeedIndex correct.
When we clear the screen after the first page that is accesssed, the start render
and SpeedIndex gets correct for the second page. The TTFB is still completely wrong
but that's ok for now because it is not used to calculate SpeedIndex. TTFB will hopefully
work when we switch to H2.

Bug: T129735
Change-Id: I237a48fd8e1b54d7c09d92421e60ead8f54c5d64