Set SpeedIndex as default metric when picking medianRun
By default WebPageTest picks the median run on pageLoadTime. But
we uses fastest run and SpeedIndex as median metrics. Lets default
to SpeedIndex (we use that in our tests today) and log
the right URL so when accessing WebPageTest we get the sorting
we want (sorted by medianRun & median metric).

Bug: T135421
Change-Id: I14f1d6e4e508ba180d7d5e2eb6f086ae3e9e6f19