Decrease wait time to 1 second
There's a bug in Chromium where connections are closed for H2
if they are idle for 5 seconds. That could be the problem that
we sometimes get really high TTFB (only for Chrome) according
to them, see:
https://bugs.chromium.org/p/chromium/issues/detail?id=700360

I'm not 100% sure but lets test to decrease the wait time
from 2 seconds to 1 second and hope we are within the span.

Bug: T153640
Change-Id: I244a6e5c1c72064bf783bcb49d6c545a883269a9