Add test for run in Tokyo and Mumbai for testing the new Asian cache pop.
These tests are meant to run hourly (or even every 4 horus) a couple of weeks
before and after we add the Asian cache pop.

All new tests are in one file with hard coded location meaning when you run
the test/rulethemall.sh you will use the same locations as we do in production.

I think that is ok for now, since we will just use them for a month or so.

About the URLs: For the Japanese wiki Japan seems to always be a top 10 URL
and the other one is a really popular pop group. For India we use one of
the most popular movies and a page about the tax change. We test on the
english Wikipedia since the traffic there from India is much higher than
for example the Hindi version.

We keep running those on Windows since we know they work.

Bug: T168416
Change-Id: Ic49f1a4422345ec5fbfaeca3b4e055e64add5a6a