Add two extra runs where we pick the median run
We have a problem with the new Linux instance that the variance
in metrics are much higher than on Windows. By default we pick the
fastest run. Let us make two other runs where we test the median
and the compare the values over time.

It will add some extra time to the runs but I really like the idea
of doing them at the same time, it will make it easier to spot
the difference.

Bug: T186582
Change-Id: I08366ef996a42a5c228ccca8aff2a7ebc696655d