Collect visual element metrics
We have turned on visual element metrics for WebPageTest and want
to test it out. At the moment we cannot configure which elements
to pick (the WPT API isn't updated) so instead of pickup the logo
we can use the generic background image pickup (on desktop).

Let us use the following when we test:
* Largest image within the viewport
* Largest heading (h1/h2) in the viewport
* Largest background image

These are the default ones. I renamed the metrics to skip the
"hero" name because that name give the elements a "hero" status
and that is not true. Better name them what they actually are.

I haven't added standard deviation for these metric yet, I will
do that in a follow up when we have collected the metrics for a
while and we are sure we wanna use them.

Bug: T202251
Change-Id: Ibd6105ee9d17701fb17f09dbd24dfd9a0355e264